@@ -4,14 +4,15 @@ version: 1.0.0
 homepage: https://github.com/applitom/flutter_net_promoter_score
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.4.4 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   
-  flutter_score_slider: ^1.0.1
-  
+  flutter_score_slider:
+    git: git@github.com:Elr03/flutter_score_slider.git
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -20,6 +21,7 @@ dev_dependencies:
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
+  #example: any
 flutter:
 
   # To add assets to your package, add an assets section, like this:
