@@ -10,24 +10,24 @@ import 'model/net_promoter_score_result.dart';
 import 'model/nps_survey_page.dart';
 
 /// Show a modal Net Promoter Score as a material design bottom sheet.
-/// 
+///
 /// The `context` argument is used to look up the [Navigator] and [Theme] for
 /// the bottom sheet. It is only used when the method is called. Its
 /// corresponding widget can be safely removed from the tree before the bottom
 /// sheet is closed.
-/// 
+///
 /// Use `onClosePressed` parameter to get callback when the user close the survery using the close button
-/// 
+///
 /// Use `onScoreChanged` parameter to get callback when the user change the score
-/// 
+///
 /// Use `onFeedbackChanged` parameter to get callback when the user change the feedback text field
-/// 
+///
 /// Use `onSurveyCompleted` parameter to get callback when the survery is campleted. This callback will provide [NetPromoterScoreResult] object with the final survery results.
-/// 
+///
 /// Use `texts` parameter to customize the text in the survery using your own texts.
-/// 
+///
 /// Use `theme` parameter to customize the look and feel of the survey. change font and colors using [ThemeData] object
-/// 
+///
 /// Use `thankYouIcon` parameter to provide a your own [Widget] for the Thank You view image
 /// ```dart
 /// showNetPromoterScore(
@@ -59,22 +59,20 @@ import 'model/nps_survey_page.dart';
 ///   theme: ThemeData.dark()
 /// );
 /// ```
-/// 
-Future<T> showNetPromoterScore<T>({
-  @required BuildContext context,
-  ThemeData theme,
-  VoidCallback onClosePressed,
-  Function(int newScore) onScoreChanged,
-  Function(String newFeedback) onFeedbackChanged,
-  Function(NetPromoterScoreResult result) onSurveyCompleted,
+///
+Future showNetPromoterScore<T>({
+  required BuildContext context,
+  ThemeData? theme,
+  VoidCallback? onClosePressed,
+  Function(int newScore)? onScoreChanged,
+  Function(String newFeedback)? onFeedbackChanged,
+  Function(NetPromoterScoreResult result)? onSurveyCompleted,
   NpsSurveyTexts texts = const NpsSurveyTexts(),
-  Widget thankYouIcon,
+  Widget? thankYouIcon,
 }) {
-  assert(texts != null);
-
   bool currentlyShowingSurvey = true;
 
-  Future<T> future = showModalBottomSheet(
+  Future future = showModalBottomSheet(
     backgroundColor: Colors.transparent,
     isDismissible: false,
     isScrollControlled: true,
@@ -83,15 +81,11 @@ Future<T> showNetPromoterScore<T>({
       return FlutterNetPromoterScore(
         onClosePressed: () {
           Navigator.pop(context);
-          if (onClosePressed != null) {
-            onClosePressed();
-          }
+          if (onClosePressed != null) onClosePressed();
         },
         onSurveyCompleted: (NetPromoterScoreResult result) {
           // call survey completion block
-          if (onSurveyCompleted != null) {
-            onSurveyCompleted(result);
-          }
+          if (onSurveyCompleted != null) onSurveyCompleted(result);
 
           // Dismiss after delay
           Future.delayed(
@@ -122,12 +116,12 @@ Future<T> showNetPromoterScore<T>({
 
 class FlutterNetPromoterScore extends StatefulWidget {
   final NpsSurveyTexts texts;
-  final VoidCallback onClosePressed;
-  final void Function(NetPromoterScoreResult result) onSurveyCompleted;
-  final Function(int newScore) onScoreChanged;
-  final Function(String newFeedback) onFeedbackChanged;
-  final ThemeData theme;
-  final Widget thankYouIcon;
+  final VoidCallback? onClosePressed;
+  final void Function(NetPromoterScoreResult result)? onSurveyCompleted;
+  final Function(int newScore)? onScoreChanged;
+  final Function(String newFeedback)? onFeedbackChanged;
+  final ThemeData? theme;
+  final Widget? thankYouIcon;
 
   FlutterNetPromoterScore({
     this.onSurveyCompleted,
@@ -137,18 +131,18 @@ class FlutterNetPromoterScore extends StatefulWidget {
     this.theme,
     this.texts = const NpsSurveyTexts(),
     this.thankYouIcon,
-  }) : assert(texts != null);
+  });
 
   @override
   FlutterNetPromoterScoreState createState() => FlutterNetPromoterScoreState();
 }
 
 class FlutterNetPromoterScoreState extends State<FlutterNetPromoterScore> {
-  int _currentScore;
+  int? _currentScore;
   String _currentFeedbackText = "";
 
   NpsSurveyPage _currentPage = NpsSurveyPage.score;
-  List<Widget Function()> _pageBuilders = List<Widget Function()>();
+  List<Widget Function()> _pageBuilders = [];
 
   @override
   void initState() {
@@ -182,18 +176,15 @@ class FlutterNetPromoterScoreState extends State<FlutterNetPromoterScore> {
       },
       onFeedbackTextChanged: (String feedbackText) {
         _currentFeedbackText = feedbackText;
-        if (this.widget.onFeedbackChanged != null) {
-          this.widget.onFeedbackChanged(feedbackText);
-        }
+        if (this.widget.onFeedbackChanged != null)
+          this.widget.onFeedbackChanged!(feedbackText);
       },
       onClosePressed: () {
-        if (this.widget.onClosePressed != null) {
-          this.widget.onClosePressed();
-        }
+        if (this.widget.onClosePressed != null) this.widget.onClosePressed!();
       },
       feedbackText: _currentFeedbackText,
       texts: this.widget.texts.feedbackPageTexts,
-      promoterType: _currentScore.toPromoterType(),
+      promoterType: _currentScore?.toPromoterType(),
     );
   }
 
@@ -204,14 +195,11 @@ class FlutterNetPromoterScoreState extends State<FlutterNetPromoterScore> {
       },
       onScoreChanged: (int score) {
         _currentScore = score;
-        if (this.widget.onScoreChanged != null) {
-          this.widget.onScoreChanged(score);
-        }
+        if (this.widget.onScoreChanged != null)
+          this.widget.onScoreChanged!(score);
       },
       onClosePressed: () {
-        if (this.widget.onClosePressed != null) {
-          this.widget.onClosePressed();
-        }
+        if (this.widget.onClosePressed != null) this.widget.onClosePressed!();
       },
       score: _currentScore,
       texts: this.widget.texts.selectScorePageTexts,
@@ -219,20 +207,18 @@ class FlutterNetPromoterScoreState extends State<FlutterNetPromoterScore> {
   }
 
   void _finilizeResult() {
-    if (this.widget.onSurveyCompleted != null) {
-      NetPromoterScoreResult finalResult = NetPromoterScoreResult();
-      finalResult.score = _currentScore;
-      finalResult.feedback = _currentFeedbackText;
-      finalResult.promoterType = _currentScore.toPromoterType();
-
-      this.widget.onSurveyCompleted(finalResult);
-    }
+    NetPromoterScoreResult finalResult = NetPromoterScoreResult();
+    finalResult.score = _currentScore;
+    finalResult.feedback = _currentFeedbackText;
+    finalResult.promoterType = (_currentScore ?? 0).toPromoterType();
+    if (this.widget.onSurveyCompleted != null)
+      this.widget.onSurveyCompleted!(finalResult);
   }
 
   @override
   Widget build(BuildContext context) {
     return Theme(
-      data: this.widget.theme == null ? Theme.of(context) : this.widget.theme,
+      data: this.widget.theme ?? Theme.of(context),
       child: SafeArea(
         child: Padding(
           padding: EdgeInsets.only(
