@@ -10,10 +10,7 @@ class NpsSelectScorePageTexts {
     this.detractorScoreLabelText = "Not likely",
     this.promoterScoreLabelText = "Very likely",
     this.submitButtonText = "SEND",
-  })  : assert(surveyQuestionText != null),
-        assert(detractorScoreLabelText != null),
-        assert(promoterScoreLabelText != null),
-        assert(submitButtonText != null);
+  });
 }
 
 class NpsFeedbackPageTexts {
@@ -38,14 +35,7 @@ class NpsFeedbackPageTexts {
         "Let us know how we can make the app better",
     this.editScoreButtonText = "Edit Score",
     this.submitButtonText = "Done",
-  })  : assert(promoterMainLabelText != null),
-        assert(promoterFeedbackTextFieldPlaceholderText != null),
-        assert(passiveMainLabelText != null),
-        assert(passiveFeedbackTextFieldPlaceholderText != null),
-        assert(detractorMainLabelText != null),
-        assert(detractorFeedbackTextFieldPlaceholderText != null),
-        assert(editScoreButtonText != null),
-        assert(submitButtonText != null);
+  });
 }
 
 class NpsThankYouPageTexts {
@@ -53,7 +43,7 @@ class NpsThankYouPageTexts {
 
   const NpsThankYouPageTexts({
     this.thankYouLabelText = "Thanks for your feedback!",
-  }) : assert(thankYouLabelText != null);
+  });
 }
 
 class NpsSurveyTexts {
@@ -65,7 +55,5 @@ class NpsSurveyTexts {
     this.selectScorePageTexts = const NpsSelectScorePageTexts(),
     this.feedbackPageTexts = const NpsFeedbackPageTexts(),
     this.thankYouPageTexts = const NpsThankYouPageTexts(),
-  })  : assert(selectScorePageTexts != null),
-        assert(feedbackPageTexts != null),
-        assert(thankYouPageTexts != null);
+  });
 }
