@@ -1,4 +1,3 @@
-
 enum PromoterType {
   promoter, // (9 =< Score <= 10)
   passive, // (7 =< Score <= 8)
@@ -8,9 +7,7 @@ enum PromoterType {
 
 extension IntToPromoterTypeExtention on int {
   PromoterType toPromoterType() {
-    if (this == null) {
-      return PromoterType.unknown;
-    } else if (this >= 0 && this <= 6) {
+    if (this >= 0 && this <= 6) {
       return PromoterType.detractor;
     } else if (this >= 7 && this <= 8) {
       return PromoterType.passive;
